requirements.txt
matplotlib==3.1.1
tensorflow==2.0.0
numpy==1.17.2
opencv_python==4.1.0.25
python-dotenv==0.10.3
